feat: add RAM options in settings modal
Underneath the java path input, there is a new range slider to allow the user to set the amount of RAM to be dedicated to Minecraft.

This data is stored in GlobalSettings under the key "java". When the game is launched, these are read from the store and passed as arguments to the argument generator.