Add more support for customizing java launch paramaters (#81)
* feat: add new store for global settings

Store interface is defined in GlobalSettingsStore.ts

Current options are for java installation

* fix(lint): fix all eslint problems

* feat(webpack): split up into multiple chunks

* feat: added padding to top of content container

* feat: export Turbolinks

* feat: add new settings modal with java options

* feat: add toast for uncaught exceptions

* fix: toast not appearing on certain exceptions

* feat: move settings button to instance list page

Modified style for new instance button and settings button

* feat: log uncaught exceptions in console

* chore: remove useless comments

* feat: add last played

* chore: remove useless code

* style: eslint

* feat: add refresh list button to instances page

* refactor: log errors in catch instead of toast function

* feat: hide error message if user is not logged in

* fix: update package.json to add missing files from webpack chunks (#77)

Changed the build configuration for electron-builder to include chunks generated by webpack. Previously, webpack only generated one chunk named renderer.bundle.js but the code has been split up into multiple chunks to improve performance.

* style: remove trailing spaces

* feat: add RAM options in settings modal

Underneath the java path input, there is a new range slider to allow the user to set the amount of RAM to be dedicated to Minecraft.

This data is stored in GlobalSettings under the key "java". When the game is launched, these are read from the store and passed as arguments to the argument generator.

* style: remove all unused imports

* style: fix indentation

* style: change spaces to tab

* style: add line at end of file

* test: add setting to add final new line when saving