Migrate to web components (#86)
* refactor: render instance list using custom elements

Beginning of migration to web components and custom elements. Created new custom element with tag "instance-list" that serves as a wrapper around the component.

* docs: add new planned feature to README.md

Added "View log output" under "Launch" as Work in Progress.

* refactor: handle renrendering event in InstanceList instead of calling a method

The InstanceList component now handles changes to the InstanceStore correctly by rerendering the list.

The function Render.instanceList() is removed.

* fix: missing files for last commit

* refactor: seperate InstanceItems to seperate web component

New custom element "InstanceItem" representing an item in the instance list.

* chore: remove useless file

* feat: auto rerendering of instance list

Removed "Refresh list" button.

* perf: move removing child elements right before rendering

* fix: focusing onto window throws error

The "this" variable was referencing the event instead of the custom component.

* fix: remove duplicates in semantic.json

* feat: add fonts for fomantic-ui to enable offline

* feat: render user status with custom element

* perf: write data to store in one shot

* feat: does not rerender user status except on login/logout or name change

* chore: remove consoleUtils

Useless and hid where the message was logged in devtools.

* chore: add Content-Security-Policy to markup

* feat: make app single instance application

It is no longer possible to open two instances of the app side by side.

* feat: prevent creation of new BrowserWindow with middle click or ctrl click

* style: eslint

* refactor: call rename modal from InstanceItem

* feat: do not rerender navbar on every load

* fix: update CSP

* refactor: rename func renameModal to rename in InstanceItem

* refactor: render instance modals from InstanceItem component

The modals delete, saves and options for instances are now all handled by the custom element InstanceItem. All events are attached when the element is created.

* chore: remove unused imports

* refactor: handle install from InstanceItem

* chore: remove useless import

* refactor: handle game launch with InstanceItem

* fix: loading page does not create a new instance of setInterval

Fixed bug causing every time a user goes onto the instance page to create a new instance of setInterval. This caused the app to render the instance list repetitively if the user loaded the instance page multiple times.

* fix: modification of ApplicationStore only triggers one render of instance list